Add 3rd dimension
Note that the TODO comment has it just plotting one of the z layers, but it does apply for all of them (I checked manually).

The W,S and LPF are not tuned yet (I had to adjust them to "extreme" values just to get something to show), so we can tune those back and get better edge/feature detection now.  I just turned them up super extreme to make sure it was working.